requirements.txt
# OpenCV
opencv

# tkinter
tkinter

# requests
requests

# hashlib
hashlib

# moviepy
moviepy

# paddleocr
fpaddleocr

# tqdm
tqdm